finishhhhhhhh

+++ src/main/java/com/ironhack/bankingSystem/model/Admin.java
@@ -3,7 +3,6 @@ package com.ironhack.bankingSystem.model;
 import com.ironhack.bankingSystem.model.security.Role;
 import com.ironhack.bankingSystem.model.security.User;
 import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
@@ -14,7 +13,7 @@ import java.util.Collection;
 @NoArgsConstructor
 public class Admin extends User {
 
-    public Admin(Long id, String name, String username, String password, Collection<Role> roles) {
+    public Admin(Integer id, String name, String username, String password, Collection<Role> roles) {
         super(id, name, username, password, roles);
     }
